Written by: Aleksander Piotrowski <alek@openbsd.org>
Import cyrus-imapd 2.2.12

The Cyrus IMAP server differs from other IMAP server implementations in
that it is generally intended to be run on sealed servers, where normal
users are not permitted to log in. The mailbox database is stored in
parts of the filesystem that are private to the Cyrus IMAP system. All
user access to mail is through the IMAP, POP3, or KPOP protocols.

From Antoine Jacoutot <ajacoutot@lphp.org>

help & ok naddy@, thanks!