Written by: Mathieu Sauve-Frankel <msf@openbsd.org>
- initial import of WWW-Mechanize-1.12

"WWW::Mechanize", or Mech for short, helps you automate interaction with
a website. It supports performing a sequence of page fetches including
following links and submitting forms. Each fetched page is parsed and
its links and forms are extracted. A link or a form can be selected,
form fields can be filled and the next page can be fetched. Mech also
stores a history of the URLs you've visited, which can be queried and
revisited.

ok alek@