Written by: Aleksander Piotrowski <alek@openbsd.org>
Import ledger 2.2

Ledger is an accounting tool with the moxie to exist. It provides no
bells or whistles, and returns the user to the days before user
interfaces were even a twinkling in their father's CRT. What it does
offer is a double-entry accounting ledger with all the flexibility and
muscle of its modern day cousins, without any of the fat. Think of it as
the Bran Muffin of accounting tools.

From Jim Razmus <jim@bonetruck.org>