Written by: Marc Espie <espie@openbsd.org>
fwbuilder is a GUI to create firewall policies from templates. It can
handle a few appliances, and PF.

Having this around may help some people start up on OpenBSD (yeah, right)
and it also comes with a few reasonable policies.

Nice things:
- the produced files look decent for mechanically generated files.
- fwbuilder can take advantage of packet-filter specific features.
For instance, it knows about modulate state, or can tweak the parameters
for state keeping on pf.