Written by: Aleksander Piotrowski <alek@openbsd.org>
Import dnsmasq 2.22

Dnsmasq is lightweight, easy to configure DNS forwarder and
DHCP server. It is designed to provide DNS and, optionally,
DHCP, to a small network. It can serve the names of local
machines which are not in the global DNS. The DHCP server
integrates with the DNS server and allows machines with
DHCP-allocated addresses to appear in the DNS with names
configured either in each host or in a central configuration
file. Dnsmasq supports static and dynamic DHCP leases and BOOTP
for network booting of diskless machines. 

From Jonathan Weiss <jw@innerewut.de>