Written by: Anil Madhavapeddy <avsm@openbsd.org>
initial import of ocaml-rss-0.3
--
OCaml-RSS is a library to parse RSS 2.0 files (which are XML files)
and build a structure representing the document.  Some functions
are also provided to print RSS documents from the structure.

The parser tries also to parse some RDF files, but many fields of
these RDF files are not taken into account.