Written by: Robert Nagy <robert@openbsd.org>
Disalble the internal mp3lib (mpg123) on arm, because it relies on the FPU.
Instead of using that let's enable the mad flavor by default on this arch.
This way you can watch movies on the Zaurus. A little description is added
to the DESCR file for people who want to use mplayer on the Zaurus.