Written by: Marc Espie <espie@openbsd.org>
conflicts from the past: history since 3.7.

As noted on ports@ recently, pkg_add -r relies on conflicts, and the
sheer existence of updates means we MUST take the past into account in
conflicts now.

Note the renaming of hugs98 to valid package names where versions are
concerned.

This commit shows clearly the renaming of the xfce4 plugin packages, the
ditching of eclipse flavors, the splitting of nessus into subpackages,
the splitting of various other software documentations, some packaging bugs
in kdeedu, and a lot of files moving around...

okay pvalchev@