Written by: Anil Madhavapeddy <avsm@openbsd.org>
initial import of cvsps-2.1
--
CVSps is a program for generating 'patchset' information from a CVS
repository. A patchset in this case is defined as a set of changes
made to a collection of files, and all committed at the same time
(using a single 'cvs commit' command). This information is valuable
to seeing the big picture of the evolution of a cvs project. While
cvs tracks revision information, it is often difficult to see what
changes were committed 'atomically' to the repository.