Written by: Marc Espie <espie@openbsd.org>
tweak _print-package-args a bit.
Add internal NODEPS to turn off computing dependencies.

Write code similar to _print-package-args that is strictly ports based,
and hence really walks the wantlibs, and needs the pkg_create stuff
to get at the real ports contents.

Recode print-package-signature to refer to the ports tree contents
exclusively (pkg_info -S for the package).

Thx bernd@ for trying that out.

This should make for a slow, but accurate, out-of-date printer.