Written by: Mathieu Sauve-Frankel <msf@openbsd.org>
- initial import of Glib2-1.101

This module provides perl access to GLib and GLib's GObject libraries.
GLib is a portability and utility library; GObject provides a generic
type system with inheritance and a powerful signal system. Together
these libraries are used as the foundation for many of the libraries
that make up the Gnome environment, and are used in many unrelated
projects.

ok sturm@