Written by: Peter Valchev <pvalchev@openbsd.org>
(same as firefox/minimo)
Fix a crash in prdtoa.c:mult() where there is an off-by-some bytes
in memory accessed during some very fucked up pointer acrobatics.
Allocate 2 bytes extra for that case, which seems to be a
sufficient hack to keep this working.

Debugged by me with minimo on zaurus/arm (pain!), seen also by krw@
and others on amd64, who helped with testing.  Due to random malloc
this of course did not show up everywhere all the time.