Written by: Aleksander Piotrowski <alek@openbsd.org>
Import potrace 1.7

potrace is a utility for tracing a bitmap, which means, transforming a
bitmap into a smooth, scalable image. The input is a bitmap (PBM, PGM,
PPM, or BMP format), and the default output is an encapsulated
PostScript file (EPS). A typical use is to create EPS files from scanned
data, such as company or university logos, handwritten notes, etc. The
resulting image is not "jaggy" like a bitmap, but smooth. It can then be
rendered at any resolution. 

From Matt Jibson <dolmant@gmail.com>