Written by: Jacob Meuser <jakemsr@openbsd.org>
the original RCS Id of readxml.c is in the patch, and when
the patch got committed to the ports tree, the Id got updated
to the ports version by cvs.  result: patch does not apply.
solution: remove the '$'s from readxml.c's RCS Id.

PKGNAME not bumped because the port has not been buildable
since the last bump.