Written by: Nikolay Sturm <sturm@openbsd.org>
Initial import of xrootconsole 0.6

xrootconsole is a small utility which displays its input in
a transparent text box on X's root window.  It will read
from any file listed on the command line or, by default,
from stdin.  It is most useful when it reads from a FIFO;
this allows you to redirect multiple commands to the FIFO
and monitor their output.

from Sebastian Trahm <inthisdefiance at gmx.net>