Written by: Kurt Miller <kurt@openbsd.org>
Disable Junk malloc option for java. java mprotects portions of the thread
stack. When forking, the child process deallocates the threads stacks it
inheritated from the parent. The junk malloc option then attempts to write
to the deallocated memory and segfaults because its not all writable. The
child proc stays in a SIGSEGV loop when this happens.  okay ian@