Written by: Marc Balmer <mbalmer@openbsd.org>
Update PostgreSQL to version 8.1.2 and move the README.OpenBSD file to the
server package, so that the informationis always available and not only
when the docs package is installed.  This was requested by many people.

SECURITY:
Version 8.1.2 fixes security and other problems that were present in 8.1.0.
See http://archives.postgresql.org/pgsql-announce/2006-01/msg00001.php for
details.