Written by: Steven Mestdagh <steven@openbsd.org>
security update to 1.4.2.1

The security issue is caused due to "gpgv" exiting with a return code
of 0 even if the detached signature file did not carry any signature.
This may result in certain scripts that use "gpgv" to conclude that
the signature is correctly verified.

More info: http://secunia.com/advisories/18845/

ok bernd@ pvalchev@