Written by: Nikolay Sturm <sturm@openbsd.org>
Initial import of HTTP-Server-Simple-0.18

This is a simple standalone HTTP server. By default, it doesn't thread or
fork. It does, however, act as a simple frontend which can be used to
build a standalone web-based application or turn a CGI into one.

from Deanna Phillips <deanna at sdf.lonestar.org>