Written by: Mathieu Sauve-Frankel <msf@openbsd.org>
initial import of Class::Accessor::Chained from submitter Sean Comeau

A chained accessor is one that always returns the object when called
with parameters (to set), and the value of the field when called with no
arguments.

This module subclasses Class::Accessor in order to provide the same
mk_accessors interface.