Written by: Robert Nagy <robert@openbsd.org>
Import of OpenOffice.Org version 2.0.3m179.  
 
OpenOffice is an Open Source, community-developed, multi-platform office
productivity suite. It includes the key desktop applications, such as a
word processor, spreadsheet, presentation manager, and drawing program,
with a user interface and feature set similar to other office suites.
 
I would like to thank everyone who helped porting:
Peter Valchev, Kurt Miller, Dale Rahn, Ian Darwin and the OpenOffice.Org team.
 
This import is an explicit request by pvalchev@ and espie@.
This means that the current ports lock is still in effect and this import
is only a planned exception by the release managers.

This port is not yet connected to the builds.
 
okay pvalchev@, espie@