Written by: Marc Espie <espie@openbsd.org>
more fixes for new issues: let the update cookie depend on the `full'
package cookie, so that an update will get all subpackages built.

And a fix for an old issue: let make update proceed to update ALL subpackages,
as seems the most natural. Create a subupdate target if one specifically
wants to update a single package (and use it as dependency in the FORCE_UPDATE
case).