Written by: Marc Balmer <mbalmer@openbsd.org>
Update to CUPS 1.2.4, which has lots of improvements over the old version
we had, like better support for foreign languages and the usual plethora
of bugfixes.

tested on various architectures and with various applications that use CUPS
as their printing system.  I verified that all ports that are depending on
CUPS still build.