Written by: Aleksander Piotrowski <alek@openbsd.org>
Import py-wxPython 2.6.3.2

wxPython is a GUI toolkit for the Python programming language. It allows
Python programmers to create programs with a robust, highly functional
graphical user interface, simply and easily. It is implemented as a
Python extension module (native code) that wraps the popular wxWidgets
cross platform GUI library, which is written in C++.