Written by: Marc Espie <espie@openbsd.org>
special case: if PKGPATH is the LIB_DEPENDS base, then it means we're
building a subpackage, and then we definitely want the `current' stuff,
not the stuff that's already installed.

Allows shared libs updates in MULTI_PACKAGES to proceed gracefully without
having to uninstall stuff.

(we do not do this stuff in general, because in other cases, builds are
`staggered', e.g., a port is built against the existing base, not the stuff
in other WRKDIRS)