Written by: Marc Balmer <mbalmer@openbsd.org>
Update to Horde 3.1.3.  This also fixes some security problems with the
previous version.

Horde now has a dependency on PostgreSQL, which is the default database.  If
you want to use MySQL instead, then use the mysql FLAVOR.  If you want to use
LDAP in addition to an SQL database, use the ldap FLAVOR.