Written by: Kurt Miller <kurt@openbsd.org>
- enable KDE integration and add openoffice-kde subpackge for this
optional feature
- enable building with java and add openoffice-java subpackage for this
optional feature. this corrects the "save as" problem and provides help
files in the base package. so now java is needed to build OOo but not to
run it. some OOo features require java to run, so install the
openoffice-java subpackage for that.
- make sbase, smath, setofficelang and spadmin applications available
via wrapper script.
- tweak configure to find versioned Qt shared libs.
- adjust include and lib search order again - sigh. put /usr/local
before jdk but after OOo dirs. this allows system libjpeg to be used
over jdk libjpeg and still avoids build conflicts with things like cppunit.
- remove extra patch to readlicense_oo/util/makefile.pmk

some testing and input ian@ okay robert@