Written by: Marc Espie <espie@openbsd.org>
belt and suspenders, best of both worlds.
bulk-builds will first try to build stuff with the java pseudo-flavor.
If that fails (because java is currently broken, or because it's not
supported on that arch), then it will build through the second line.

If the packages already exist, the infrastructure will notice, and the
second build line will do nothing.

consensus reached after discussion with nikolay, kurt, pval...