Written by: Marc Espie <espie@openbsd.org>
start of multi-packages simplification: if MULTI_PACKAGES contains -main,
consider it to be the main package, and do some equal treatment for each
package.

(to do: add subpackage where needed to WANTLIB, RUN_DEPENDS, LIB_DEPENDS
and act on them accordingly).

Also define _DONE_FILES for _fetch-makefile to avoid blocking if it's
not invoked from a higher level target.