Written by: Marc Espie <espie@openbsd.org>
big internal changes: if MULTI_PACKAGES is not defined, define it to '-',
so that the `normal' cases is MULTI_PACKAGES, (with possibly one
special '-' subpackage).

Adjust a few tests accordingly so that people don't notice the '-'
SUBPACKAGE: mostly do not print some directory changes, adjust FULLPKGPATH,
and set up PLIST- templates correctly.