Written by: Marc Espie <espie@openbsd.org>
Add proper "Returning to build of..." to unconfuse the logger in the case
of FORCE_UPDATE.   Also break out in case update fails (otherwise the old
installed package would be enough to continue building).

Sprinkle REPORT_PROBLEM to know more precisely what stuff is breaking.

replace some
if eval cmd; then
with
if (eval exec cmd); then

because if eval... seems to break early, even with the test guard, thus
giving a chance to REPORT_PROBLEM to do stuff instead of having to wait.

(Note: a trap would probably be simpler ?)