Written by: Deanna Phillips <deanna@openbsd.org>
Bringing this in now so that we can work on it here.  Batting tarballs
back and forth in email messages doesn't make any sense when we have a
cvs tree to work in.

So, not hooked into the package builds yet.

Original work and port name from dlg.

ok robert@, dlg@