Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import py-Xlib 0.13

The Python X Library is intended to be a fully functional X client
library for Python programs. It is written entirely in Python, in
contrast to earlier X libraries for Python (the ancient X extension and
the newer plxlib) which were interfaces to the C Xlib.

ok ajacoutot@