Written by: Steven Mestdagh <steven@openbsd.org>
import gnucap 0.35

Gnu Circuit Analysis Package

The primary component is a general purpose circuit simulator. It
performs nonlinear dc and transient analyses, fourier analysis, and ac
analysis. It is fully interactive and command driven. It can also be run
in batch mode or as a server. Spice compatible models for the MOSFET
(level 1-7), BJT, and diode are included in this release.

submitted by andreas.bihlmaier at gmx.de