Written by: Robert Nagy <robert@openbsd.org>
import of btpd-0.12:

btpd is a bittorrent  client consisting of a daemon and client commands,        
which can be used to read and/or manipulate the daemon state. The daemon        
is capable of running several torrents simultaneously and only uses one         
tcp port. It's fairly low on resource usage and should be perfect for           
file distribution sites. Efficient downloads and ease of use makes this         
client a good choice for the casual user as well

from Martin Cronier <hex29a@openbsd.se>
fixes from Vlad Glagolev <stelzy@gmail.com> and me