Written by: Robert Nagy <robert@openbsd.org>
import of iso-codes-0.58;

This package aims to provide a central list of the country, language,
and currency names and their translations.  The data comes from
ISO standards 639 (language names), 3166 (country names), and 4217 (currency).
The translations are from the Translation Project. The data is provided
in tabular files as well as XML.