Written by: David Gwynne <dlg@openbsd.org>
switch apache2 from a gnu style configure to a simple one so we can
define prefix the way apache likes it.

this is because apache2 has a different understanding of what the prefix
means, and our understanding and application of it on this port leads to
extremely confused paths in a lot of its generated files.

our understanding of prefix is to mean the path at which the binaries,
libs, manpages, and so on are stored, ie, /usr/local. apache2 understands
prefix to mean "install architecture-independent files", or in real terms
the ServerRoot. obviously using /usr/local as the server root when we want
to use /var/apache2 for that purpose is uncomfortable for it, and it leads
to things like broken paths in the default config files and builds of
modules.

ok robert@