Written by: David Gwynne <dlg@openbsd.org>
import of mod_perl 2.0.3 for use on apache 2

this package aims to provide the full power of perl together with the
apache web server version 2. you can use perl to manage apache, respond to
requests for web pages and much more.

there already exists a mod_perl in ports (www/mod_perl) for use with the
apache in the base openbsd system. this one is usable only with
www/apache-httpd. the mod_perl developers were kind enough to set things up
so they wont conflict with each other, so you can have both running
concurrently.

originally from srebrenko sehic, with some naming tweaks from me. lots of
help from robert@ (as always).

im importing this now so it can be worked on in tree, but not hooking it up
to the bulk builds. there are a few warts that need to be fixed but i hate
sending tarballs around when we have a tree to develop in. its also nice to
have commit messages against changes so you can see why decisions were made
a certain way. after everyone is happy it can be hooked up.