Written by: Kurt Miller <kurt@openbsd.org>
Update to patchset 4 which includes a security fix for CVE-2007-0243

Buffer overflow allows applets to gain privileges via a GIF image with a
block with a 0 width field, which triggers memory corruption and possible
execution of arbitrary code.

Also included in patchset 4 are corrections for building the jdk on amd64
systems with 4G+ memory and the inclusion of the mozilla headers needed to
build the plugin without an external dependency on the old mozilla suite.