Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import gxmessage-2.6.2

gxmessage is a GTK2 based xmessage clone trying to be as compatible as
possible.  You might like it if you're running a mostly GTK desktop.
gxmessage opens a window to display a message obtained from the command
line, from a file, or from stdin.  The window includes a row of buttons,
each of which causes the program to exit with a different return code.

"looks fine" steven@