Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import gcompris-8.2.2

GCompris ("I got it") is an educational software which propose different
activities to children from 2 to 10 years old.
Some of the activities are game orientated, but nonetheless still
educational.  Below you can find a list of categories with some of the
activities available in that category.
* computer discovery: keyboard, mouse, different mouse gesture
* algebra: table memory, enumeration, double entry table, mirror image
* science: the canal lock, the water cycle, the submarine, electric
  simulation
* geography: place the country on the map
* games: chess, memory, connect 4, oware, sudoku
* reading: reading practice
* other: learn to tell time, puzzle of famous paintings, vector drawing,
  cartoon making, ...

Currently GCompris offers in excess of 100 activities and more are being
developed.

"slap it in" mbalmer@