Written by: Mathieu Sauve-Frankel <msf@openbsd.org>
initial import of barcode 0.98

barcode is meant to solve most needs in barcode creation with a
conventional printer. It can create printouts for the conventional
product tagging standards: UPC-A, UPC-E, EAN-13, EAN-8, ISBN, as well as
a few other formats. Ouput is generated as either Postscript or
Encapsulated Postscript (other back-ends may be added if needed).

ok pvalchev@