Written by: Marc Espie <espie@openbsd.org>
tweak the modules so they also define MODQTn_* values along the usual
MODQT_* values.

This makes it possible to include both qt3 and qt4, as long as one is
very careful in choosing the right variables, without any change for the
existing qt3-only, qt4-only ports.

We don't bother with qt and qt2 as they are mostly end-of-life.