Written by: Kurt Miller <kurt@openbsd.org>
update to 1.2.21. This version addresses the following security flaw which
effects versions 1.2.19 and 1.2.20 only:

CVE-2007-0774 A Long URL Stack Overflow Vulnerability exists in the URI
handler for the mod_jk library. When parsing a long URL request, the URI
worker map routine performs an unsafe memory copy. This results in a stack
overflow condition which can be leveraged execute arbitrary code.
http://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2007-0774